backend/requirements.txt
bcrypt
passlib

flask
flask-script
flask-migrate
flask-cors

marshmallow
marshmallow-sqlalchemy

psycopg2

raven[flask]
webargs
sqlalchemy-utcdatetime

twopi-flask-utils[celery]
git+http://github.com/nickw444/flask-sqlalchemy.git

flask-classy
flask-login

requests-oauthlib
requests

flask-wtf